refactor(scrt-contract): refactor macro into 7 files
